fix: address settings center review feedback
