+++ src/constants/code-block.ts
@@ -1,0 +1,1 @@
+export const COLLAPSE_LINE_THRESHOLD = 8;
