+++ src/styles/global/motion.css
@@ -13,7 +13,8 @@ html.motion-off .lightrope li,
 html.motion-off .animate-ping,
 html.motion-off .sider-slot.entering,
 html.motion-off .sider-slot.exiting {
-  animation: none !important;
+  /* html.motion-off … outranks the single-class animation definitions, so no !important needed. */
+  animation: none;
 }
 
 @media (prefers-reduced-motion: reduce) {
@@ -27,6 +28,7 @@ html.motion-off .sider-slot.exiting {
   .animate-ping,
   .sider-slot.entering,
   .sider-slot.exiting {
+    /* biome-ignore lint/complexity/noImportantStyles: same specificity as the source animations, which are imported after this file, so !important is required to win. */
     animation: none !important;
   }
 }
